4893 - extract table logic into hook and move footer inside the table
